Handle unsigned conversion warning

+++ lib/ngtcp2_pkt.c
@@ -539,7 +539,8 @@ ngtcp2_ssize ngtcp2_pkt_decode_frame(ngtcp2_frame *dest, const uint8_t *payload,
     return ngtcp2_pkt_decode_datagram_frame(&dest->datagram, payload,
                                             payloadlen);
   default:
-    if ((type & ~(NGTCP2_FRAME_STREAM - 1)) == NGTCP2_FRAME_STREAM) {
+    if ((type & ~((uint64_t)NGTCP2_FRAME_STREAM - 1)) ==
+        (uint64_t)NGTCP2_FRAME_STREAM) {
       return ngtcp2_pkt_decode_stream_frame(&dest->stream, payload, payloadlen);
     }
